Nest Git under VersionControlSystems namespace
